ADDED PRODUCT domain model encompases batch Moved batch, allocate method here
